+++ src/yuni/ui/gl/drawingsurface.h
@@ -31,7 +31,7 @@ namespace UI
 
 	public:
 		//! Constructor to an empty transparent surface
-		DrawingSurface(uint width, uint height);
+		DrawingSurface(float width, float height);
 
 		//! Destructor
 		~DrawingSurface();
@@ -49,32 +49,32 @@ namespace UI
 		void clear();
 
 		//! Resize the surface
-		void resize(uint width, uint height);
+		void resize(float width, float height);
 
 		//! Draw text on the surface
 		void drawText(const String& text, const FTFont::Ptr& font, const Color::RGBA<float>& color,
-			int x, int y);
+			float x, float y);
 
 		//! Draw text on the surface, centered in a rectangle
 		void drawTextInRect(const String& text, const FTFont::Ptr& font,
-			const Color::RGBA<float>& color, int x, int y, uint width, uint height);
+			const Color::RGBA<float>& color, float x, float y, float width, float height);
 
 		//! Draw an anti-aliased line on the surface
-		void drawLine(const Color::RGBA<float>& color, int startX, int startY, int endX, int endY,
-			float lineWidth);
+		void drawLine(const Color::RGBA<float>& color, float startX, float startY,
+			float endX, float endY, float lineWidth);
 
 		//! Draw a line anti-aliased with a given background color
 		void drawLine(const Color::RGBA<float>& color, const Color::RGBA<float>& bgColor,
-			int startX, int startY, int endX, int endY, float lineWidth);
+			float startX, float startY, float endX, float endY, float lineWidth);
 
 		//! Draw an empty rectangle (borders only)
 		void drawRectangle(const Color::RGBA<float>& frontColor,
-			const Color::RGBA<float>& backColor, int x, int y, uint width, uint height,
+			const Color::RGBA<float>& backColor, float x, float y, float width, float height,
 			float lineWidth);
 
 		//! Draw a rectangle filled with a background color on the surface
 		void drawFilledRectangle(const Color::RGBA<float>& frontColor,
-			const Color::RGBA<float>& backColor, int x, int y, uint width, uint height,
+			const Color::RGBA<float>& backColor, float x, float y, float width, float height,
 			float lineWidth);
 
 		/*!
@@ -82,15 +82,16 @@ namespace UI
 		**
 		** \warning `imageOpacity` is used only for the image, the alpha from `fillColor` is used for out of bounds opacity
 		*/
-		void drawImage(const Gfx3D::Texture::Ptr& texture, int x, int y, uint width, uint height,
-			const Color::RGBA<float>& fillColor, DisplayMode dispMode = dmNone,
-			int offsetX = 0, int offsetY = 0, float imageOpacity = 1.0f);
+		void drawImage(const Gfx3D::Texture::Ptr& texture, float x, float y,
+			float width, float height, const Color::RGBA<float>& fillColor,
+			DisplayMode dispMode = dmNone,
+			float offsetX = 0, float offsetY = 0, float imageOpacity = 1.0f);
 
 		//! Fill with a color. Use clipping to control the filled area
 		void fill(const Color::RGBA<float>& color);
 
 		//! Activate clipping inside a rectangle
-		void beginRectangleClipping(int x, int y, uint width, uint height);
+		void beginRectangleClipping(float x, float y, float width, float height);
 
 		//! Disable last clipping, must be called in pair with begin*Clipping
 		void endClipping();
@@ -99,10 +100,10 @@ namespace UI
 		const Gfx3D::Texture::Ptr& texture() const;
 
 		//! Get the current width of the surface
-		uint width() const;
+		float width() const;
 
 		//! Get the current height of the surface
-		uint height() const;
+		float height() const;
 
 	private:
 		//! Update the texture according to the changes in the surface
